+++ CHANGELOG.md
@@ -4,10 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- proper require of dry-container dependency in OperationsCenter
+
 ## [0.0.1] - 2018-11-18
 ### Added
 - basic gem scaffold
 - test infrastructure
 - implemented /ping endpoint
 
+[Unreleased]: https://github.com/ncreuschling/dsapi/commits/master
 [0.0.1]: https://github.com/ncreuschling/dsapi/releases/tag/v0.0.1